requirements.txt
PySide6>=6.5.0
pdf2docx>=0.5.6
pytesseract>=0.3.10
Pillow>=10.0.0
darkdetect>=0.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
